% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPSAQ.R
\name{batch_psaq_analysis}
\alias{batch_psaq_analysis}
\title{Perform Batch PSAQ Analysis on Multiple Excel Files}
\usage{
batch_psaq_analysis(dir, xlsx_files_pattern)
}
\arguments{
\item{dir}{A character string specifying the directory where the Excel files must be searched.
These sequences are used to generate theoretical isotopic distributions.}

\item{xlsx_files_pattern}{A character string specifying the pattern to match Excel files.
Example pattern : "^784.*\\\.xlsx$", which matches files starting with "784"
and ending with ".xlsx". This pattern is passed to \code{\link{list.files}}.}
}
\value{
This function does not return a value. Instead, it writes the results of the deconvolution
        to new Excel files in the current working directory. Each output file is named
        "calculated_ratios_[original_filename]".
}
\description{
Perform batch processing of PSAQ (Protein Standard Absolute Quantification) analysis
on multiple Excel files matching a specific pattern. This function reads each file,
deconvolutes peptide abundances, and saves the results to new Excel files.
}
\details{
This function performs the following steps for each Excel file matching the pattern:
1. Identifies all Excel files in the current working directory that match the specified pattern.
2. For each file, reads the file and call \code{\link{deconvoluate_peptides_abundances}} to process the data.
3. Writes the results to a new Excel file with the prefix "calculated_ratios_" followed by the original filename.

The function uses \code{\link{deconvoluate_peptides_abundances}} to deconvolute peptide abundances
by separating endogenous and PSAQ contributions from MS2 isotopic data.
}
\examples{
# Perform batch PSAQ analysis on files starting with "784" and ending with ".xlsx"
batch_psaq_analysis("inst/extdata", "^784.*\\\\.xlsx$")

# Perform batch PSAQ analysis on files starting with "sample" and ending with ".xlsx"
batch_psaq_analysis("inst/extdata", "^sample.*\\\\.xlsx$")

}
\seealso{
\code{\link{deconvoluate_peptides_abundances}} for deconvoluting peptide abundances from a single Excel file.
\code{\link{list.files}} for listing files matching a pattern.
}
