% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopes_deconv.R
\name{compute_all_fragments_distributions}
\alias{compute_all_fragments_distributions}
\title{Generate Isotopic Distributions of Fragments from Peptide Sequences}
\usage{
compute_all_fragments_distributions(sequences, fragment_type = c("y"))
}
\arguments{
\item{sequences}{A character vector containing one or more peptide sequences
(e.g., c("ALQASALAAWGGK", "PEPTIDEK")).}

\item{fragment_type}{A character vector specifying the types of fragments to generate.
Possible values are "y", "b", or c("y", "b") (default is "y").
If "y", y-ion fragments are generated. If "b", b-ion fragments are generated.
If both are specified, both y-ion and b-ion fragments are generated.}
}
\value{
A data frame containing the combined theoretical fragments and their isotopic distributions
        for all input sequences. The data frame includes the following columns:
        \describe{
          \item{id}{Integer representing the fragment number.}
          \item{precursor_seq}{Character string of the precursor peptide sequence.}
          \item{precursor_mass}{Double representing the monoisotopic mass of the precursor peptide.}
          \item{fragment_label}{Character string representing the label of the fragment (e.g., "y3", "b5").}
          \item{fragment_seq}{Character string of the fragment sequence.}
          \item{C, H, N, O, S}{Integers representing the number of each type of atom in the fragment.}
          \item{fragment_mass}{Double representing the monoisotopic mass of the fragment.}
          \item{ms1_isotopic_ratio}{Double representing the isotopic ratio (M+1/M) at the MS1 level.}
          \item{ms2_isotopic_ratio}{Double representing the isotopic ratio (M+1/M) at the MS2 level for each fragment.}
        }
}
\description{
Generate theoretical fragments (y and/or b ions) and their isotopic distributions
for a list of peptide sequences.
}
\details{
This function iterates through a list of peptide sequences and generates theoretical fragments
with their isotopic distributions for each sequence. It combines the results into a single data frame.
The function uses \code{\link{compute_fragments_isotopic_distribution}} to calculate the fragments
and their isotopic distributions for each individual sequence.
}
\examples{
# Generate y-ion fragments and their isotopic distributions for multiple peptides
compute_all_fragments_distributions(c("ALQASALAAWGGK", "PEPTIDEK"))

# Generate both y-ion and b-ion fragments for multiple peptides
compute_all_fragments_distributions(c("ALQASALAAWGGK", "PEPTIDEK"), fragment_type = c("y", "b"))

}
\seealso{
\code{\link{compute_fragments_isotopic_distribution}} for computing fragments and isotopic distributions
of a single peptide sequence.
}
