% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopes_deconv.R
\name{generate_peptide_fragments}
\alias{generate_peptide_fragments}
\title{Enumerate y and/or b fragments of a peptide}
\usage{
generate_peptide_fragments(sequence, fragment_type = c("y"))
}
\arguments{
\item{sequence}{A character string representing the amino acid sequence of the peptide (e.g., `"ALQASALAAWGGK"`).}

\item{fragment_type}{A character vector specifying the types of fragments to generate.
Possible values are `"y"`, `"b"`, or `c("y", "b")` (default is `"y"`).
If `"y"`, y-ion fragments are generated. If `"b"`, b-ion fragments are generated.
If both are specified, both y-ion and b-ion fragments are generated.}
}
\value{
A data frame with the following columns:
        \describe{
          \item{id}{Integer representing the fragment number.}
          \item{precursor_seq}{Character string of the precursor peptide sequence.}
          \item{precursor_mass}{Double representing the monoisotopic mass of the precursor peptide.}
          \item{fragment_label}{Character string representing the label of the fragment (e.g., `"y3"`, `"b5"`).}
          \item{fragment_seq}{Character string of the fragment sequence.}
          \item{C}{Integer representing the number of carbon atoms in the fragment.}
          \item{H}{Integer representing the number of hydrogen atoms in the fragment.}
          \item{N}{Integer representing the number of nitrogen atoms in the fragment.}
          \item{O}{Integer representing the number of oxygen atoms in the fragment.}
          \item{S}{Integer representing the number of sulfur atoms in the fragment.}
          \item{fragment_mass}{Double representing the monoisotopic mass of the fragment.}
        }
}
\description{
This function enumerates y and/or b fragments of a peptide sequence, names each fragment,
computes their theoretical mass and atomic composition.
}
\details{
The function uses the `ConvertPeptide` function to convert peptide sequences into their elemental compositions
and the `MonoisotopicMass` function to compute the monoisotopic mass of each fragment.
The precursor mass is computed with a charge of 2, while the fragment masses are computed with a charge of 1.
The function does not consider IAA (iodoacetamide) modification.
}
\examples{
# Generate y-ion fragments for a peptide
generate_peptide_fragments("ALQASALAAWGGK")

# Generate both y-ion and b-ion fragments for a peptide
generate_peptide_fragments("ALQASALAAWGGK", fragment_type = c("y", "b"))

}
\seealso{
\code{\link[OrgMassSpecR]{ConvertPeptide}} for converting peptide sequences to elemental compositions.
\code{\link[OrgMassSpecR]{MonoisotopicMass}} for computing monoisotopic masses.
}
