% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopes_deconv.R
\name{isotopes_abundances}
\alias{isotopes_abundances}
\title{The Atoms Natural Light and Heavy Isotopes}
\usage{
isotopes_abundances()
}
\value{
a named list of isotopic distribution of light and heavy (light+1) isotopes
}
\description{
The Atoms Natural Light and Heavy Isotopes
}
