% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopes_deconv.R
\name{isotopic_distribution}
\alias{isotopic_distribution}
\title{Calculate the theoretical isotopic distribution of the first two isotopes of a peptide}
\usage{
isotopic_distribution(sequence)
}
\arguments{
\item{sequence}{A character string representing the amino acid sequence of the peptide (e.g., "ALQASALAAWGGK").}
}
\value{
A data frame with two columns:
        \describe{
          \item{proba}{Raw probabilities of each isotopologue.}
        }
}
\description{
This function computes the theoretical isotopic distribution of a precursor from its amino acids sequence.
 The distribution represents the relative abundance of isotopologues (molecules with different isotopic compositions)
 by taking into account the natural abundance of +1 isotopes (^13C, ^2H, ^15N, ^17O, and ^33S).
}
\details{
The function uses the \code{\link[OrgMassSpecR]{ConvertPeptide}} function to convert the peptide sequence
into its elemental composition, and the \code{\link[sinib]{dsinib}} function to compute the isotopic distribution.
}
\examples{
# Example usage
isotopic_distribution("ALQASALAAWGGK")
isotopic_distribution("GILAADESVGTMGNR")
isotopic_distribution("LSFSYGR")

}
\seealso{
\code{\link[OrgMassSpecR]{ConvertPeptide}} for converting peptide sequences to elemental compositions.
}
