% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPSAQ.R
\name{plot_psaq_ratios}
\alias{plot_psaq_ratios}
\title{Plot PSAQ Experimental Ratios}
\usage{
plot_psaq_ratios(psaq_deconvolution)
}
\arguments{
\item{psaq_deconvolution}{A data frame containing deconvoluted peptide abundance data.
This data frame should include the following columns:
\describe{
  \item{experimental_ratio_area}{Experimental ratio of PSAQ to endogenous absed on area abundances.}
  \item{experimental_ratio_intensity}{Experimental ratio of PSAQ to endogenous based on intensity abundances.}
  \item{seq}{Peptide sequence for coloring points by sequence.}
}}
}
\value{
A ggplot object displaying experimental PSAQ ratios. The plot includes:
        \describe{
          \item{Circular points}{Representing experimental_ratio_area values.}
          \item{X-shaped points}{Representing experimental_ratio_intensity values (if present).}
        }
        The y-axis is limited to the range [-0.5, 2.5].
}
\description{
Creates a scatter plot to visualize experimental PSAQ (Protein Standard Absolute Quantification) ratios
from deconvoluted peptide abundance data.
}
\details{
This function creates a scatter plot where:
\enumerate{
  \item The x-axis represents the row indices of the input data frame.
  \item The y-axis represents the experimental ratio values.
  \item Points are colored by peptide sequence.
  \item Circular points represent experimental_ratio_area values.
  \item X-shaped points represent experimental_ratio_intensity values (if this column exists in the data).
  \item The y-axis is constrained between -0.5 and 2.5 to focus on relevant ratio values.
}
}
\examples{
# Deconvolute then plot PSAQ example data
psaq_data = deconvolute_peptides_abundances(sample_10ng_r4)
plot = plot_psaq_ratios(psaq_data)


}
\seealso{
\code{\link{deconvolute_peptides_abundances}} for generating the input data.
\code{\link[ggplot2]{ggplot}} for creating the base plot.
\code{\link[ggplot2]{geom_point}} for adding points to the plot.
}
