% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_sample_10ng_r4.R
\docType{data}
\name{sample_10ng_r4}
\alias{sample_10ng_r4}
\title{sample_10ng_r4}
\format{
A data frame with 114 rows and 11 variables:
\describe{
  \item{ seq }{  character }
  \item{ mz }{  numeric }
  \item{ frag_mz }{  numeric }
  \item{ rt }{  numeric }
  \item{ m/z }{  numeric }
  \item{ frag_isotope }{  character }
  \item{ frag_label }{  character }
  \item{ status }{  numeric }
  \item{ Xpl1_007843_auto.mzdb.intensity }{  numeric }
  \item{ Xpl1_007843_auto.mzdb.scans_count }{  numeric }
  \item{ Xpl1_007843_auto.mzdb.area }{  numeric }
}
}
\source{
Source
}
\usage{
sample_10ng_r4
}
\description{
Description.
}
\keyword{datasets}
