% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopes_deconv.R
\name{compute_fragments_isotopic_distribution}
\alias{compute_fragments_isotopic_distribution}
\title{Compute the Isotopic Distribution of MS2 Fragments of a Precursor}
\usage{
compute_fragments_isotopic_distribution(sequence, fragment_type = c("y"))
}
\arguments{
\item{sequence}{A character string representing the amino acid sequence of the peptide (e.g., "ALQASALAAWGGK").}

\item{fragment_type}{A character vector specifying the types of fragments to generate.
Possible values are "y", "b", or c("y", "b") (default is "y").
If "y", y-ion fragments are generated. If "b", b-ion fragments are generated.
If both are specified, both y-ion and b-ion fragments are generated.}
}
\value{
A data frame containing the following columns:
        \describe{
          \item{id}{Integer representing the fragment number.}
          \item{precursor_seq}{Character string of the precursor peptide sequence.}
          \item{precursor_mass}{Double representing the monoisotopic mass of the precursor peptide.}
          \item{fragment_label}{Character string representing the label of the fragment (e.g., "y3", "b5").}
          \item{fragment_seq}{Character string of the fragment sequence.}
          \item{C, H, N, O, S}{Integers representing the number of each type of atom in the fragment.}
          \item{fragment_mass}{Double representing the monoisotopic mass of the fragment.}
          \item{ms1_isotopic_ratio}{Double representing the isotopic ratio (M+1/M) at the MS1 level.}
          \item{ms2_isotopic_ratio}{Double representing the isotopic ratio (M+1/M) at the MS2 level for each fragment.}
        }
}
\description{
Compute the isotopic distribution of y (and/or b) fragments of the specified precursor peptide sequence.
This function calculates the isotopic ratios for both MS1 and MS2 levels and appends them to the fragment data.
}
\details{
This function performs the following steps:
1. Generates the specified fragments (y and/or b ions) using the `generate_peptide_fragments` function.
2. Computes the isotopic distribution of the precursor peptide using the `isotopic_distribution` function.
3. Calculates the MS1 isotopic ratio (M+1/M) and adds it to the fragment data.
4. For each fragment, computes the MS2 isotopic distribution using the `compute_ms2_isotopic_distribution` function.
5. Calculates the MS2 isotopic ratio (M+1/M) for each fragment and appends it to the fragment data.

The isotopic ratio is defined as the ratio of the probability of the first isotopologue (M+1) to the monoisotopic peak (M).
}
\examples{
# Compute the isotopic distribution for y-ion fragments of a peptide
compute_fragments_isotopic_distribution("ALQASALAAWGGK", c("y"))

# Compute the isotopic distribution for b-ion fragments of a peptide
compute_fragments_isotopic_distribution("ALQASALAAWGGK", c("b"))

}
\seealso{
\code{\link{generate_peptide_fragments}} for generating peptide fragments.
\code{\link{isotopic_distribution}} for computing the isotopic distribution of a peptide.
\code{\link{compute_ms2_isotopic_distribution}} for computing the isotopic distribution of MS2 fragments.
}
