% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPSAQ.R
\name{deconvolute_peptides_abundances}
\alias{deconvolute_peptides_abundances}
\title{Deconvolute Peptide Abundances from MS2 Data}
\usage{
deconvolute_peptides_abundances(ms2_data)
}
\arguments{
\item{ms2_data}{A data frame containing MS2 fragments first and second isotopes abundances.}
}
\value{
A data frame containing deconvoluted peptide abundances with the following additional columns:
        \describe{
          \item{i1_endo_intensity}{Calculated intensity of the endogenous second isotope.}
          \item{I0_psaq_intensity}{Calculated intensity of the PSAQ first isotope.}
          \item{I0_endo_intensity}{Calculated intensity of the endogenous first isotope.}
          \item{experimental_ratio_intensity}{Ratio of PSAQ to endogenous intensities.}
          \item{i1_endo_area}{Calculated area of the endogenous second isotope.}
          \item{I0_psaq_area}{Calculated area of the PSAQ first isotope.}
          \item{I0_endo_area}{Calculated area of the endogenous first isotope.}
          \item{experimental_ratio_area}{Ratio of PSAQ to endogenous areas.}
        }
        The function returns NA for infinite values.
}
\description{
Deconvolute peptide abundances by separating endogenous and PSAQ (Protein Standard Absolute Quantification)
contributions from MS2 isotopic data. This function processes experimental data to calculate
intensities and areas for endogenous and PSAQ peptides.
}
\details{
This function performs the following steps:
1. Validates the required columns.
2. Extracts and renames columns for first and second isotope data.
3. Joins first and second isotope data.
4. Generates theoretical isotopic distributions for the provided sequences.
5. Merges experimental data with theoretical distributions.
6. Calculates corrected intensities and areas for endogenous and PSAQ peptides.
7. Computes experimental ratios for both intensities and areas.

The input data frame must contain the following columns:
\describe{
  \item{seq}{Peptide sequence.}
  \item{mz}{Precursor m/z.}
  \item{frag_mz}{Fragment m/z.}
  \item{rt}{Retention time.}
  \item{frag_isotope}{Isotope label ("first" or "second").}
  \item{frag_label}{Fragment label (e.g., "y3").}
  \item{intensity}{One or more columns with "intensity" in their names.}
  \item{area}{One or more columns with "area" in their names.}
}
}
\examples{
# Deconvolute peptide abundances from MS2 data
data = deconvolute_peptides_abundances(sample_10ng_r4)

}
\seealso{
\code{\link{compute_all_fragments_distributions}} for generating theoretical fragment distributions.
\code{\link{reshape_data}} for reshaping abundance.
}
