% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopes_deconv.R
\name{isotopic_distribution}
\alias{isotopic_distribution}
\title{Calculate the theoretical isotopic distribution of a peptide}
\usage{
isotopic_distribution(sequence, size = 4)
}
\arguments{
\item{sequence}{A character string representing the amino acid sequence of the peptide (e.g., "ALQASALAAWGGK").}

\item{size}{An integer specifying the number of isotopologues to compute (default = 4).
This corresponds to the number of isotope peaks to model, starting from the monoisotopic peak.}
}
\value{
A data frame with two columns:
        \describe{
          \item{proba}{Raw probabilities of each isotopologue.}
          \item{percent}{Relative abundance of each isotopologue as a percentage of the monoisotopic peak.}
        }
}
\description{
This function computes the theoretical isotopic distribution of a precursor from its amino acids sequence.
 The distribution represents the relative abundance of isotopologues (molecules with different isotopic compositions)
 by taking into account the natural abundance  of heavy isotopes (^13C, ^2H, ^15N, ^18O, and ^33S).
It is useful for predicting isotope patterns in mass spectrometry.
}
\details{
The function uses the \code{\link[OrgMassSpecR]{ConvertPeptide}} function to convert the peptide sequence
into its elemental composition, and the \code{\link[sinib]{dsinib}} function to compute the isotopic distribution.
}
\examples{
# Example usage
isotopic_distribution("ALQASALAAWGGK", size = 4)

}
\seealso{
\code{\link[OrgMassSpecR]{ConvertPeptide}} for converting peptide sequences to elemental compositions.
\code{\link[sinib]{dsinib}} for computing isotopic distributions.
}
