% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopes_deconv.R
\name{ms2_isotopic_distribution}
\alias{ms2_isotopic_distribution}
\title{Compute the Isotopic Distribution of the first two isotopes of a Fragment}
\usage{
ms2_isotopic_distribution(fragment_seq, precursor_seq)
}
\arguments{
\item{fragment_seq}{A character string representing the amino acid sequence of the fragment.}

\item{precursor_seq}{A character string representing the amino acid sequence of the precursor peptide.}
}
\value{
A data frame with the following columns:
        \describe{
          \item{proba}{A numeric vector containing the probability associated with each MS2 isotope.}
        }
        The probabilities of each isotope is based on the number of heavy atoms included in the fragment.
}
\description{
Compute the isotopic distribution of the specified fragment, taking into account that only
the second isotope of the precursor have been included in the selection window.
}
\examples{
# Compute the isotopic distribution for a fragment of the precursor peptide
ms2_isotopic_distribution("LAAWGGK", "ALQASALAAWGGK")
ms2_isotopic_distribution("WGGK", "ALQASALAAWGGK")
ms2_isotopic_distribution("ALQASALAAWGGK", "ALQASALAAWGGK")
ms2_isotopic_distribution("SYGR", "LSFSYGR")

}
\seealso{
\code{\link[OrgMassSpecR]{ConvertPeptide}} for converting peptide sequences to elemental compositions.
}
